Force slide level to 2
The slide level usually is the highest header level in the hierarchy that is
followed immediately by content.
This means that if you accidentally put text on the first level:

 # Header
 asdf
 ## Subheader

Then pandoc changes the layout scheme.
Forcing level=2 prevents the layout from accidentally getting messed up
(but does not show the "asdf" content)